Move the adapter to its parent package
