+++ app/src/main/java/com/yurii/youtubemusic/screens/saved/mediaitems/MediaListAdapter.kt
@@ -1,4 +1,4 @@
-package com.yurii.youtubemusic.adapters
+package com.yurii.youtubemusic.screens.saved.mediaitems
 
 import android.content.Context
 import android.support.v4.media.session.PlaybackStateCompat
